Finish restructuring projects into their own packages
Now compiles and all that jazz
