+++ PokeAutobuilder/Pages/PokemonTeamStoragePage.razor
@@ -2,9 +2,10 @@
 
 @page "/teamstorage"
 
+@using PokemonDataModel
+@using PokeAutobuilder.Source.Services;
+@using PokeAutobuilder.Shared.Cards
 @using PokeApiNet
-@using pokeAutoBuilder.Shared.Cards
-@using pokeAutoBuilder.Source.Services;
 
 <PageTitle>Team Storage</PageTitle>
