+++ PokeAutobuilder/Pages/TeamBuilderPage.razor
@@ -1,6 +1,7 @@
 ﻿@using PokeApiNet
-@using pokeAutoBuilder.Source.Services;
-@using pokeAutoBuilder.Shared;
+@using PokeAutobuilder.Source.Services;
+@using PokeAutobuilder.Shared;
+@using PokemonDataModel
 
 @inject SessionService Session
 @inject ProfileService Profile
