+++ PokeAutobuilder/Shared/AutoBuilderDialog.razor
@@ -1,7 +1,9 @@
-﻿@using pokeAutoBuilder.Shared.Panes
-@using pokeAutoBuilder.Source.Services;
-@using pokeAutoBuilder.Source.TeamGeneration;
-@using pokeAutoBuilder.Shared.Cards
+﻿@using AutoBuilder
+@using PokeAutobuilder.Shared.Cards
+@using PokeAutobuilder.Shared.Panes
+@using PokeAutobuilder.Source.Services
+@using PokemonDataModel
+@using Utility
 
 @inject SessionService Session
 @inject ProfileService Profile
