+++ PokeAutobuilder/Shared/Cards/PokemonMiniCard.razor
@@ -1,7 +1,9 @@
 ﻿@inject ProfileService Profile
 @inject IDialogService DialogService
 
-@using pokeAutoBuilder.Source.Services;
+@using PokeAutobuilder.Source.Services
+@using PokemonDataModel
+@using Utility
 
 @if (DeleteFromStorageMode)
 {
