+++ PokeAutobuilder/Shared/Cards/PokemonTeamCard.razor
@@ -3,7 +3,9 @@
 @inject IDialogService DialogService
 @inject NavigationManager NavigationManager
 
-@using pokeAutoBuilder.Source.Services;
+@using PokeAutobuilder.Source.Services
+@using PokemonDataModel
+@using Utility
 
 <MudStack Spacing="1" AlignItems="AlignItems.Start" Class="ma-1">	
 	<MudStack Row="true" AlignItems="AlignItems.End" Class="mud-width-full">
