+++ PokeAutobuilder/Shared/Panes/PokemonDefensePane.razor
@@ -1,4 +1,7 @@
-﻿@if (Pokemon is not null)
+﻿@using PokemonDataModel
+@using Utility
+
+@if (Pokemon is not null)
 {
 	<MudPaper Class="d-flex flex-column align-center justify-center pa-2">
 		<MudStack Spacing="5" AlignItems="AlignItems.Center" Style="width: 100%;" Class="pa-2">
