+++ PokeAutobuilder/Shared/Panes/PokemonStatsChartPane.razor
@@ -1,4 +1,7 @@
-﻿@if (Pokemon is not null)
+﻿@using PokemonDataModel
+@using Utility
+
+@if (Pokemon is not null)
 {
 	<MudPaper Class="d-flex flex-column align-center justify-center mud-width-full pa-2">
 		<MudText Typo="Typo.h6">Base Stats</MudText>
