+++ PokeAutobuilder/Shared/Panes/PokemonTeamCoveragePane.razor
@@ -1,4 +1,7 @@
-﻿<MudPaper Class="mud-width-full pa-2 overflow-x-auto">
+﻿@using PokemonDataModel
+@using Utility
+
+<MudPaper Class="mud-width-full pa-2 overflow-x-auto">
 	<MudText Typo="Typo.h6" Align="Align.Center">Coverage</MudText>
 	<MudSimpleTable Style="width: 100%;" Dense="true" Elevation="0">
 		<thead>
