+++ PokeAutobuilder/Shared/PokemonDetailsDialog.razor
@@ -1,9 +1,10 @@
 ﻿@inject PokeApiService PokeApiService
 
+@using PokemonDataModel
+@using PokeAutobuilder.Source.Services;
+@using PokeAutobuilder.Shared.Panes
+@using PokeAutobuilder.Shared.Cards
 @using PokeApiNet;
-@using pokeAutoBuilder.Shared.Cards
-@using pokeAutoBuilder.Shared.Panes
-@using pokeAutoBuilder.Source.Services;
 
 <MudDialog>
     <DialogContent>
