+++ PokeAutobuilder/Shared/PokemonSearchBox.razor
@@ -1,8 +1,11 @@
-﻿@using PokeApiNet
+﻿@using PokemonDataModel
+@using PokeAutobuilder.Source.Services;
+@using PokeAutobuilder.Shared.Cards
+@using PokeApiNet
+@using Utility
+
+@using static PokeAutobuilder.Pages.TeamBuilderPage
 @using Type = PokeApiNet.Type
-@using pokeAutoBuilder.Shared.Cards
-@using pokeAutoBuilder.Source.Services;
-@using static pokeAutoBuilder.Pages.TeamBuilderPage
 
 @inject ProfileService Profile
 @inject SessionService Session
