+++ PokeAutobuilder/Shared/PokemonTeamDetailsDialog.razor
@@ -1,5 +1,6 @@
-﻿@using pokeAutoBuilder.Shared.Cards
-@using pokeAutoBuilder.Shared.Panes
+﻿@using PokeAutobuilder.Shared.Cards
+@using PokeAutobuilder.Shared.Panes
+@using PokemonDataModel
 
 @if (Team is not null)
 {
