+++ PokeAutobuilder/Shared/StatsPanel.razor
@@ -1,5 +1,7 @@
-﻿@using Type = PokeApiNet.Type
-@using pokeAutoBuilder.Shared.Panes
+﻿@using PokeAutobuilder.Shared.Panes
+@using PokemonDataModel
+
+@using Type = PokeApiNet.Type
 
 @if (Pokemon is not null)
 {
