+++ PokeAutobuilder/Shared/StatsTeamPanel.razor
@@ -1,5 +1,7 @@
-﻿@using Type = PokeApiNet.Type
-@using pokeAutoBuilder.Shared.Panes
+﻿@using PokeAutobuilder.Shared.Panes
+@using PokemonDataModel
+
+@using Type = PokeApiNet.Type
 
 <MudPaper Style="background: var(--mud-palette-background)" Elevation="4">
 	<MudGrid Class="pa-4" Justify="Justify.Center">
