+++ PokeAutobuilder/Shared/StorageBox.razor
@@ -1,5 +1,6 @@
-﻿@using pokeAutoBuilder.Shared.Cards
-@using pokeAutoBuilder.Source.Services;
+﻿@using PokeAutobuilder.Shared.Cards
+@using PokeAutobuilder.Source.Services;
+@using PokemonDataModel
 
 @inject ProfileService Profile
