+++ PokeAutobuilder/_Imports.razor
@@ -7,7 +7,7 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
-@using pokeAutoBuilder
-@using pokeAutoBuilder.Shared
-@using pokeAutoBuilder.Source
+@using PokeAutobuilder
+@using PokeAutobuilder.Shared
+@using PokeAutobuilder.Source
 @using MudBlazor
